# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Sales Outreach Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign for selected prospects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign detail with prospects
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Edit campaign fields (subject/body/etc) before sending
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/generate:
    post:
      operationId: generateCampaignEmail
      tags: [campaigns]
      summary: Generate subject/body draft with Gemini
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Generated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/send-test:
    post:
      operationId: sendTestEmail
      tags: [campaigns]
      summary: Send a test email to the account owner
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Test email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Send campaign immediately to pending prospects (respecting daily cap)
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignSendResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/approve-template:
    post:
      operationId: approveCampaignTemplate
      tags: [campaigns]
      summary: One-time approval required before the agent can autonomously send this campaign's template
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Template approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "400":
          description: Template not generated yet
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/schedule:
    post:
      operationId: scheduleCampaign
      tags: [campaigns]
      summary: Schedule campaign for future sending
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignScheduleInput"
      responses:
        "200":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------
  # Email inbox / threads
  # ---------------------------------------------------------------------
...
components:
  parameters:
...
    CampaignId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    ActionResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }

    # ------------------------ Prospects ------------------------
...
    CampaignStatus:
      type: string
      enum: [draft, scheduled, sending, sent, completed]
    CampaignProspectStatus:
      type: string
      enum: [pending, sent, replied, bounced, stopped]
    Campaign:
      type: object
      required:
        [
          id,
          name,
          goal,
          tone,
          productDescription,
          targetAudience,
          cta,
          status,
          followupCount,
          templateApproved,
          prospectCount,
          sentCount,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        goal: { type: string }
        tone: { type: string }
        productDescription: { type: string }
        targetAudience: { type: string }
        cta: { type: string }
        subject: { type: ["string", "null"] }
        body: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/CampaignStatus" }
        followupCount: { type: integer }
        templateApproved: { type: boolean }
        scheduledAt: { type: ["string", "null"], format: date-time }
        sentAt: { type: ["string", "null"], format: date-time }
        prospectCount: { type: integer }
        sentCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CampaignProspect:
      type: object
      required: [id, prospectId, companyName, status, followupStage, createdAt]
      properties:
        id: { type: integer }
        prospectId: { type: integer }
        companyName: { type: string }
        contactEmail: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/CampaignProspectStatus" }
        followupStage: { type: integer }
        lastEmailAt: { type: ["string", "null"], format: date-time }
        nextFollowupAt: { type: ["string", "null"], format: date-time }
        gmailThreadId: { type: ["string", "null"] }
        stoppedReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CampaignDetail:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          required: [prospects]
          properties:
            prospects:
              type: array
              items: { $ref: "#/components/schemas/CampaignProspect" }
    CampaignInput:
      type: object
      required:
        [name, goal, tone, productDescription, targetAudience, cta, prospectIds]
      properties:
        name: { type: string, minLength: 1 }
        goal: { type: string, minLength: 1 }
        tone: { type: string, minLength: 1 }
        productDescription: { type: string, minLength: 1 }
        targetAudience: { type: string, minLength: 1 }
        cta: { type: string, minLength: 1 }
        prospectIds:
          type: array
          items: { type: integer }
          minItems: 1
    CampaignUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        goal: { type: string }
        tone: { type: string }
        productDescription: { type: string }
        targetAudience: { type: string }
        cta: { type: string }
        subject: { type: string }
        body: { type: string }
    CampaignScheduleInput:
      type: object
      required: [scheduledAt]
      properties:
        scheduledAt: { type: string, format: date-time }
    CampaignSendResult:
      type: object
      required: [sent, queued, failed, suppressed, campaign]
      properties:
        sent: { type: integer }
        queued: { type: integer }
        failed: { type: integer }
        suppressed: { type: integer }
        campaign: { $ref: "#/components/schemas/CampaignDetail" }

    # ------------------------ Threads / inbox ------------------------
...